+++ CHANGELOG.md
@@ -5,6 +5,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+  - Python's built-in `in` keyword now works with a ConfigFile.
+
+    Example:
+    ```python
+    config = ConfigFile('./pyproject.toml')
+
+    'tool.poetry' in config
+    >>> True
+    ```
+
+### Changed
+
+ - Depreciated `stringify()` in favor of just using the built-in `str()`.
+
+### Fixed
+
+ - Addresses issue #25 (INI parser isn't converting back to string).
+
 ## 0.11.0 - 2020-08-07
 
 ### Added